evaluation/requirements.txt
boto3>=1.34.0
google-cloud-translate>=3.15.0
sacrebleu>=2.4.0